// bandhanmidea/src/app/model/Education.model.ts
export interface Education {
  userId: number;
  method: string;
  highestDegree: string;

  sscPY: string;
  sscGroup: string;
  sscResult: string;
  sscInsName: string;

  hscPY: string;
  hscGroup: string;
  hscResult: string;
  hscInsName: string;

  ugPY: string;
  gSubject: string;
  ugInsName: string;

  otherDetails: string;
}